+++ app/scripts/TrackRenderer.js
@@ -1289,6 +1289,7 @@ class TrackRenderer extends React.Component {
       this.valueScaleZooming = false;
       this.element.__zoom = this.zoomStartTransform;
     }
+
     this.props.pubSub.publish('app.zoomEnd');
   }
 
@@ -1494,11 +1495,17 @@ class TrackRenderer extends React.Component {
       scene: this.pStage,
       dataConfig,
       dataFetcher,
+      getLockGroupExtrema: () => {
+        return this.currentProps.getLockGroupExtrema(track.uid);
+      },
       handleTilesetInfoReceived,
       animate: () => {
         this.currentProps.onNewTilesLoaded(track.uid);
       },
       svgElement: this.svgElement,
+      isValueScaleLocked: () => {
+        return this.currentProps.isValueScaleLocked(track.uid);
+      },
       onValueScaleChanged: () => {
         this.currentProps.onValueScaleChanged(track.uid);
       },
@@ -1518,6 +1525,16 @@ class TrackRenderer extends React.Component {
 
     if (track.x) {
       context.xPosition = track.x;
+    }
+
+    // for viewport-projection-horizontal and viewport-projection-center
+    if (track.projectionXDomain) {
+      context.projectionXDomain = track.projectionXDomain;
+    }
+
+    // for viewport-projection-vertical and viewport-projection-center
+    if (track.projectionYDomain) {
+      context.projectionYDomain = track.projectionYDomain;
     }
 
     const options = track.options;